feat(BUX-313): removal of unused files and methods, extracting server logic to utils
