--- transports/p2p/p2putil/block.go
@@ -1,140 +1,0 @@
-// Copyright (c) 2013-2016 The btcsuite developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-package p2putil
-
-import (
-	"bytes"
-	"io"
-
-	"github.com/bitcoin-sv/block-headers-service/internal/chaincfg/chainhash"
-	"github.com/bitcoin-sv/block-headers-service/internal/wire"
-)
-
-// OutOfRangeError describes an error due to accessing an element that is out
-// of range.
-type OutOfRangeError string
-
-// BlockHeightUnknown is the value returned for a block height that is unknown.
-// This is typically because the block has not been inserted into the main chain
-// yet.
-const BlockHeightUnknown = int32(-1)
-
-// Error satisfies the error interface and prints human-readable errors.
-func (e OutOfRangeError) Error() string {
-	return string(e)
-}
-
-// Block defines a bitcoin block that provides easier and more efficient
-// manipulation of raw blocks.  It also memoizes hashes for the block and its
-// transactions on their first access so subsequent accesses don't have to
-// repeat the relatively expensive hashing operations.
-type Block struct {
-	msgBlock        *wire.MsgBlock  // Underlying MsgBlock
-	serializedBlock []byte          // Serialized bytes for the block
-	blockHash       *chainhash.Hash // Cached block hash
-	blockHeight     int32           // Height in the main block chain
-}
-
-// MsgBlock returns the underlying wire.MsgBlock for the Block.
-func (b *Block) MsgBlock() *wire.MsgBlock {
-	// Return the cached block.
-	return b.msgBlock
-}
-
-// Bytes returns the serialized bytes for the Block.  This is equivalent to
-// calling Serialize on the underlying wire.MsgBlock, however it caches the
-// result so subsequent calls are more efficient.
-func (b *Block) Bytes() ([]byte, error) {
-	// Return the cached serialized bytes if it has already been generated.
-	if len(b.serializedBlock) != 0 {
-		return b.serializedBlock, nil
-	}
-
-	// Serialize the MsgBlock.
-	w := bytes.NewBuffer(make([]byte, 0, b.msgBlock.SerializeSize()))
-	err := b.msgBlock.Serialize(w)
-	if err != nil {
-		return nil, err
-	}
-	serializedBlock := w.Bytes()
-
-	// Cache the serialized bytes and return them.
-	b.serializedBlock = serializedBlock
-	return serializedBlock, nil
-}
-
-// Hash returns the block identifier hash for the Block.  This is equivalent to
-// calling BlockHash on the underlying wire.MsgBlock, however it caches the
-// result so subsequent calls are more efficient.
-func (b *Block) Hash() *chainhash.Hash {
-	// Return the cached block hash if it has already been generated.
-	if b.blockHash != nil {
-		return b.blockHash
-	}
-
-	// Cache the block hash and return it.
-	hash := b.msgBlock.BlockHash()
-	b.blockHash = &hash
-	return &hash
-}
-
-// Height returns the saved height of the block in the block chain.  This value
-// will be BlockHeightUnknown if it hasn't already explicitly been set.
-func (b *Block) Height() int32 {
-	return b.blockHeight
-}
-
-// SetHeight sets the height of the block in the block chain.
-func (b *Block) SetHeight(height int32) {
-	b.blockHeight = height
-}
-
-// NewBlock returns a new instance of a bitcoin block given an underlying
-// wire.MsgBlock.  See Block.
-func NewBlock(msgBlock *wire.MsgBlock) *Block {
-	return &Block{
-		msgBlock:    msgBlock,
-		blockHeight: BlockHeightUnknown,
-	}
-}
-
-// NewBlockFromBytes returns a new instance of a bitcoin block given the
-// serialized bytes.  See Block.
-func NewBlockFromBytes(serializedBlock []byte) (*Block, error) {
-	br := bytes.NewReader(serializedBlock)
-	b, err := NewBlockFromReader(br)
-	if err != nil {
-		return nil, err
-	}
-	b.serializedBlock = serializedBlock
-	return b, nil
-}
-
-// NewBlockFromReader returns a new instance of a bitcoin block given a
-// Reader to deserialize the block.  See Block.
-func NewBlockFromReader(r io.Reader) (*Block, error) {
-	// Deserialize the bytes into a MsgBlock.
-	var msgBlock wire.MsgBlock
-	err := msgBlock.Deserialize(r)
-	if err != nil {
-		return nil, err
-	}
-
-	b := Block{
-		msgBlock:    &msgBlock,
-		blockHeight: BlockHeightUnknown,
-	}
-	return &b, nil
-}
-
-// NewBlockFromBlockAndBytes returns a new instance of a bitcoin block given
-// an underlying wire.MsgBlock and the serialized bytes for it.  See Block.
-func NewBlockFromBlockAndBytes(msgBlock *wire.MsgBlock, serializedBlock []byte) *Block {
-	return &Block{
-		msgBlock:        msgBlock,
-		serializedBlock: serializedBlock,
-		blockHeight:     BlockHeightUnknown,
-	}
-}
